fixing logout
