+++ app/src/main/java/com/example/final_project_android_admin/ui/airplane/AirplaneFragment.kt
@@ -17,7 +17,9 @@ import com.example.final_project_android_admin.data.api.service.ApiHelper
 import com.example.final_project_android_admin.databinding.FragmentAirplaneBinding
 import com.example.final_project_android_admin.utils.UserDataStoreManager
 import com.example.final_project_android_admin.viewmodel.AirplaneViewModel
+import com.example.final_project_android_admin.viewmodel.LoginViewModel
 import com.example.final_project_android_admin.viewmodel.factory.AirplaneViewModelFactory
+import com.example.final_project_android_admin.viewmodel.factory.UserViewModelFactory
 import com.google.android.material.snackbar.Snackbar
 
 class AirplaneFragment : Fragment(), AirplaneAdapter.ListAirplaneInterface {
@@ -25,6 +27,7 @@ class AirplaneFragment : Fragment(), AirplaneAdapter.ListAirplaneInterface {
     private val binding get() = _binding!!
 
     private lateinit var airplaneViewModel: AirplaneViewModel
+    private lateinit var viewModel: LoginViewModel
     private lateinit var pref: UserDataStoreManager
 
     override fun onCreateView(
@@ -32,7 +35,11 @@ class AirplaneFragment : Fragment(), AirplaneAdapter.ListAirplaneInterface {
         savedInstanceState: Bundle?
     ): View {
         // Inflate the layout for this fragment
+
         pref = UserDataStoreManager(requireContext())
+        viewModel = ViewModelProvider(
+            this, UserViewModelFactory(ApiHelper(ApiClient.instance), pref)
+        )[LoginViewModel::class.java]
         airplaneViewModel = ViewModelProvider(
             this, AirplaneViewModelFactory(ApiHelper(ApiClient.instance), pref)
         )[AirplaneViewModel::class.java]
@@ -122,6 +129,15 @@ class AirplaneFragment : Fragment(), AirplaneAdapter.ListAirplaneInterface {
                 R.id.transaction -> {
                     findNavController().navigate(R.id.transactionFragment)
                     true
+                }
+                R.id.logout -> {
+                    viewModel.removeIsLoginStatus()
+                    viewModel.removeId()
+                    viewModel.removeUsername()
+                    viewModel.removeToken()
+                    viewModel.getDataStoreIsLogin().observe(viewLifecycleOwner) {
+                        findNavController().navigate(R.id.loginFragment)
+                    }
                 }
                 else -> false
             }
