+++ app/src/main/java/com/example/final_project_android_admin/ui/flight/FlightFragment.kt
@@ -15,8 +15,11 @@ import com.example.final_project_android_admin.data.api.response.flight.DataFlig
 import com.example.final_project_android_admin.data.api.service.ApiClient
 import com.example.final_project_android_admin.data.api.service.ApiHelper
 import com.example.final_project_android_admin.databinding.FragmentFlightBinding
+import com.example.final_project_android_admin.utils.UserDataStoreManager
 import com.example.final_project_android_admin.viewmodel.FlightViewModel
+import com.example.final_project_android_admin.viewmodel.LoginViewModel
 import com.example.final_project_android_admin.viewmodel.factory.FlightViewModelFactory
+import com.example.final_project_android_admin.viewmodel.factory.UserViewModelFactory
 import com.google.android.material.snackbar.Snackbar
 
 class FlightFragment : Fragment(), FlightAdapter.ListFlightInterface {
@@ -24,12 +27,19 @@ class FlightFragment : Fragment(), FlightAdapter.ListFlightInterface {
     private val binding get() = _binding!!
 
     private lateinit var flightViewModel: FlightViewModel
+    private lateinit var viewModel: LoginViewModel
+    private lateinit var pref: UserDataStoreManager
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
         // Inflate the layout for this fragment
+
+        pref = UserDataStoreManager(requireContext())
+        viewModel = ViewModelProvider(
+            this, UserViewModelFactory(ApiHelper(ApiClient.instance), pref)
+        )[LoginViewModel::class.java]
 
         flightViewModel = ViewModelProvider(
             this, FlightViewModelFactory(ApiHelper(ApiClient.instance))
@@ -105,6 +115,15 @@ class FlightFragment : Fragment(), FlightAdapter.ListFlightInterface {
                 R.id.transaction -> {
                     findNavController().navigate(R.id.transactionFragment)
                     true
+                }
+                R.id.logout -> {
+                    viewModel.removeIsLoginStatus()
+                    viewModel.removeId()
+                    viewModel.removeUsername()
+                    viewModel.removeToken()
+                    viewModel.getDataStoreIsLogin().observe(viewLifecycleOwner) {
+                        findNavController().navigate(R.id.loginFragment)
+                    }
                 }
                 else -> false
             }
